+++ AA2_water_color_sort_Eloi_Ivan/AA2_water_color_sort_Eloi_Ivan/jugada.h
@@ -4,4 +4,4 @@
 void pedirSource(short& indiceSource);
 void pedirDestino(short& indiceDestino, short indiceSource);
 void moverContenido(short indiceSource, short indiceDestino, char tablero[ALTO][ANCHO]);
-short calcularPuntuacion(char tablero[ALTO][ANCHO], short intentos);
+int calcularPuntuacion(char tablero[ALTO][ANCHO], short intentos);
